Refactoring MCIS source code
