+++ src/core/mcis/orchestration.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// package to manage multi-cloud infra service
+// Package mcis is to manage multi-cloud infra service
 package mcis
 
 import (
