+++ src/core/mcis/recommendation.go
@@ -11,10 +11,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// package to manage multi-cloud infra service
+// Package mcis is to manage multi-cloud infra service
 package mcis
 
 import (
+	"encoding/json"
 	"fmt"
 	"math"
 	"sort"
@@ -23,6 +24,8 @@ import (
 
 	"github.com/cloud-barista/cb-tumblebug/src/core/common"
 	"github.com/cloud-barista/cb-tumblebug/src/core/mcir"
+
+	cbstore_utils "github.com/cloud-barista/cb-store/utils"
 )
 
 // DeploymentPlan is struct for .
@@ -327,4 +330,107 @@ func getHaversineDistance(a1 float64, b1 float64, a2 float64, b2 float64) (dista
 
 	earthRadius := float64(6371)
 	return (earthRadius * c)
+}
+
+// GetRecommendList is func to get recommendation list
+func GetRecommendList(nsId string, cpuSize string, memSize string, diskSize string) ([]TbVmPriority, error) {
+
+	fmt.Println("GetRecommendList")
+
+	var content struct {
+		Id             string
+		Price          string
+		ConnectionName string
+	}
+	//fmt.Println("[Get MCISs")
+	key := common.GenMcisKey(nsId, "", "") + "/cpuSize/" + cpuSize + "/memSize/" + memSize + "/diskSize/" + diskSize
+	fmt.Println(key)
+	keyValue, err := common.CBStore.GetList(key, true)
+	keyValue = cbstore_utils.GetChildList(keyValue, key)
+	if err != nil {
+		common.CBLog.Error(err)
+		return []TbVmPriority{}, err
+	}
+
+	var vmPriorityList []TbVmPriority
+
+	for cnt, v := range keyValue {
+		fmt.Println("getRecommendList1: " + v.Key)
+		err = json.Unmarshal([]byte(v.Value), &content)
+		if err != nil {
+			common.CBLog.Error(err)
+			return []TbVmPriority{}, err
+		}
+
+		content2 := mcir.TbSpecInfo{}
+		key2 := common.GenResourceKey(nsId, common.StrSpec, content.Id)
+
+		keyValue2, err := common.CBStore.Get(key2)
+		if err != nil {
+			common.CBLog.Error(err)
+			return []TbVmPriority{}, err
+		}
+		json.Unmarshal([]byte(keyValue2.Value), &content2)
+		content2.Id = content.Id
+
+		vmPriorityTmp := TbVmPriority{}
+		vmPriorityTmp.Priority = strconv.Itoa(cnt)
+		vmPriorityTmp.VmSpec = content2
+		vmPriorityList = append(vmPriorityList, vmPriorityTmp)
+	}
+
+	fmt.Println("===============================================")
+	return vmPriorityList, err
+
+	//requires error handling
+
+}
+
+// CorePostMcisRecommend is func to command to all VMs in MCIS with SSH
+func CorePostMcisRecommend(nsId string, req *McisRecommendReq) ([]TbVmRecommendInfo, error) {
+
+	err := common.CheckString(nsId)
+	if err != nil {
+		common.CBLog.Error(err)
+		return nil, err
+	}
+
+	/*
+		var content struct {
+			//VmReq          []TbVmRecommendReq    `json:"vmReq"`
+			VmRecommend    []mcis.TbVmRecommendInfo `json:"vmRecommend"`
+			PlacementAlgo  string                   `json:"placementAlgo"`
+			PlacementParam []common.KeyValue        `json:"placementParam"`
+		}
+	*/
+	//content := RestPostMcisRecommendResponse{}
+	//content.VmReq = req.VmReq
+	//content.PlacementAlgo = req.PlacementAlgo
+	//content.PlacementParam = req.PlacementParam
+
+	VmRecommend := []TbVmRecommendInfo{}
+
+	vmList := req.VmReq
+
+	for i, v := range vmList {
+		vmTmp := TbVmRecommendInfo{}
+		//vmTmp.RequestName = v.RequestName
+		vmTmp.VmReq = req.VmReq[i]
+		vmTmp.PlacementAlgo = v.PlacementAlgo
+		vmTmp.PlacementParam = v.PlacementParam
+
+		var err error
+		vmTmp.VmPriority, err = GetRecommendList(nsId, v.VcpuSize, v.MemorySize, v.DiskSize)
+
+		if err != nil {
+			common.CBLog.Error(err)
+			//mapA := map[string]string{"message": "Failed to recommend MCIS"}
+			//return c.JSON(http.StatusFailedDependency, &mapA)
+			return nil, fmt.Errorf("Failed to recommend MCIS")
+		}
+
+		VmRecommend = append(VmRecommend, vmTmp)
+	}
+
+	return VmRecommend, nil
 }
